refactor(benchmark): extract tag creation popover
- move benchmark tag creation into its own component

- use danger styling for the remove case action
